upload and check folder empty

+++ API/app/controllers/upload.py
@@ -8,15 +8,14 @@ upFile = UploadSet('upFile',extensions=AllExcept(()))
 
 @upload.route('', methods=['GET','POST'])
 def uploadFile():
+  emptyFolder()
   if request.method == 'POST' and 'videoFile'  in request.files:
-    emptyFolder()
     splitName=request.files['videoFile'].filename.split('.')
     request.files['videoFile'].filename='srcVideo.'+splitName[len(splitName)-1]
     filename = upFile.save(request.files['videoFile'])
     # file_url = upFile.url(filename)
     return 'don'
   elif request.method == 'POST' and 'imageFile'  in request.files:
-    emptyFolder()
     splitName=request.files['imageFile'].filename.split('.')
     request.files['imageFile'].filename='srcImage.'+splitName[len(splitName)-1]
     filename = upFile.save(request.files['imageFile'])
@@ -25,14 +24,15 @@ def uploadFile():
   else:
     return 'hi'
 
-
 # clear all file in folder
 def emptyFolder():
   folder = 'app/static'
-  for the_file in os.listdir(folder):
-      file_path = os.path.join(folder, the_file)
-      try:
-          if os.path.isfile(file_path):
-              os.unlink(file_path)
-      except Exception as e:
-          print(e)
+  print(len(os.listdir(folder)))
+  if len(os.listdir(folder))>=2:
+    for the_file in os.listdir(folder):
+        file_path = os.path.join(folder, the_file)
+        try:
+            if os.path.isfile(file_path):
+                os.unlink(file_path)
+        except Exception as e:
+            print(e)
